requirements.txt
Flask==3.0
psycopg2-binary==2.9.6
flask_cors
flask_jwt_extended
gunicorn
flask-limiter>=3.8.0